Repositories, User builder, Conference id changed to Long type

+++ src/main/java/pl/it/conf/scheduler/model/Conference.java
@@ -15,7 +15,7 @@ public class Conference {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer conferenceId;
+    private Long conferenceId;
 
     @NonNull
     private Date date;
